Save the metadata/gamelist before launching a game
There is an issue when editing metadata that if you shutdown you will lose the
metadata changes. This will save them before launching every game. It is a hack
but should fix this edge case. It uses es_settings.cfg and SaveGamelistOnLaunch
setting to turn this on and off in case it does not scale.